Master Thesis : Airbnb  Copenhagen
An AI & Machine Learning approach to analyzing and modeling Copenhagen Airbnb prices
  The most important takeaways from InsideAirbnb data show that the Gradient Boosting Regression model has a higher and more accurate estimate of the price variable
with R2 score of 0.77 and RMSE of DKK 252. In addition, Gradient Boosting has the lowest Average Error of 155 DKK and shows that the best predictions were delivered according to this
Models results in research show that on a panel of Machine Learning and Deep Learning models can construct a method and value for Airbnb learning brands and enable them to identify and predict a most accurate and performance-capable Airbnb listing price.
In this study, I modeled Airbnb listings data in Copenhagen from September 2018 to September 2019. About 80% of listings are apartments, with and the average nightly rate of 1114DKK.
based on 24 features indicated, on tree-based models, namely gradient-boosting  regression and extreme gradient-boosting  regression, explains the price variation in the training dataset quite well (pictures at coefficient of variation R2). The R2 measurement for the test dataset was fairly strong (about 0.61) with a root mean square error of about $ 327. Room_type function was a function of the greatest importance.
Airbnb data paves the way for future research in the field and proves that it is possible to identify and predict an Airbnb listing price with a field of machine learning and deep learning methods and techniques that provide an identifiable pattern for the results in this dissertation. : 
di2. This study has identified
how well a panel of six machine learning methods can predict the Airbnb price in Copenhagen and establish a basis for a sustainable model approach for Copenhagen's Airbnb environment. 
With these questions in mind, a comprehensive study of the dataset has been performed to explore the research foundations and find appropriate patterns of choice of features to assess the accuracy of machine learning models available for implementation. 
The dataset was explored to identify missing values and errors in  features of the dataset, followed by a selection process of a set of realistic features Identification patterns for the data set's missing values, cleaning process, and feature selection served as our largest data processing foundation to establish a dataset that matched the capabilities of all the ml models